round 1 of claude's suggested changes
