+++ .gitignore
@@ -7,3 +7,4 @@ temp
 __pycache__
 .idea/
 .DS_Store
+exports/
