+++ requirements-dev.txt
@@ -1,0 +1,4 @@
+-r requirements.txt
+
+streamlit
+streamlit-shortcuts
